backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
workos==5.42.0
PyJWT[crypto]>=2.0.0
httpx==0.28.1
deploy-box-python @ file:///shared/deploy-box-python
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pyperclip>=1.8.0
